refactor: extract CompactionTelemetryRecorder from LcmContextEngine
Phase 3 of the engine.ts decomposition (#867). Telemetry recording
(cache-state telemetry from runtime prompt-cache snapshots, leaf success
marking, deferred-debt rows) moves to src/compaction-telemetry.ts. The
deferred-debt drain/consume loop stays in the engine: it is compaction
orchestration wired through the session queue and executeCompactionCore.
normalizeOptionalCount joins value-utils as a free function.

Full suite: 1279 tests passing.
